We like our shots unsalted thank you very much, better for the heart. Use Shot Lists on the client. Send proper shot IDs from the server. Have the API death event have subclasses for each death type. Convert kills with no shots into smites.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_99_shot_branch@17087 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfs.h
@@ -83,7 +83,7 @@ extern void searchFlag(GameKeeper::Player &playerData);
 extern void relayPlayerPacket(int index, uint16_t len, const void *rawbuf, uint16_t code);
 extern void playerAlive(int playerIndex);
 extern int  lookupPlayer(const PlayerId& id);
-extern bool playerKilled(int victimIndex, BlowedUpReason reason, int16_t id, bool respawnOnBase = false);
+extern bool playerKilled(int victimIndex, BlowedUpReason reason, int id, bool respawnOnBase = false);
 extern bool smitePlayer(int victimIndex, BlowedUpReason reason = GotKilledMsg, bool respawnOnBase = false);
 extern void dropPlayerFlag(GameKeeper::Player &playerData, const float dropPos[3]);
 extern void captureFlag(int playerIndex, TeamColor teamCaptured);
